% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr.data.R
\docType{data}
\name{scr.data}
\alias{scr.data}
\title{Spatial Capture-Recapture test/example data}
\format{A list containing examples of data used in and generated by SCR methods.}
\usage{
data(scr.data)
}
\description{
A dataset containing examples of simple SCR capture histories,
an acoustic capture history, trap co-ordinates, and mask point co-ordinates.
}
\examples{
\dontrun{
## Data are generated as follows.
##  - The same seed is used for each call to scr.sim()
# set.seed(291112)
# scr.data$capt = scr.sim(5, 75, scr.data$traps)
# scr.data$bin.capt = scr.sim(0.8, 75, scr.data$traps, distr = "binom")
# scr.data$acoustic.capt = scr.sim(0.8, 75, scr.data$traps, acoustic = TRUE, lambda_c = 2)

data(scr.data)
scr.fit(scr.data$capt, scr.data$traps, scr.data$mask)
}
}
\keyword{datasets}
